Allow cast to u8

+++ crates/vm/src/format.rs
@@ -54,16 +54,20 @@ unsafe fn parse_grouping(grouping: *const libc::c_char) -> Vec<u8> {
     if grouping.is_null() {
         return result;
     }
+
     unsafe {
         let mut ptr = grouping;
         while ![0, libc::c_char::MAX].contains(&*ptr) {
-            result.push(*ptr);
+            #[allow(clippy::unnecessary_cast, reason = "ptr is not u8 on all platforms")]
+            result.push(*ptr as u8);
             ptr = ptr.add(1);
         }
     }
+
     if !result.is_empty() {
         result.push(0);
     }
+
     result
 }
